backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.3
xgboost>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
